System: Update llama.cpp to b8121 and enable ARM optimizations
Update llama.cpp submodule from b7692 to b8121. Enable KleidiAI
optimized kernels and OpenMP for arm64-v8a builds. Upgrade CMake
from 3.22.1 to 3.31.6 (required for KleidiAI FetchContent).
Remove outdated -std=c++11 flag and add GGML_NATIVE=OFF for
cross-compilation.

+++ app/src/main/cpp/native-lib.cpp
@@ -28,7 +28,7 @@
 class AndroidLogBuf : public std::streambuf {
 protected:
     std::streamsize xsputn(const char* s, std::streamsize n) override {
-        __android_log_print(ANDROID_LOG_INFO, "Llama", "%.*s", n, s);
+        __android_log_print(ANDROID_LOG_INFO, "Llama", "%.*s", (int)n, s);
         return n;
     }
